Move Report class to a new file
