+++ src/ui/main.py
@@ -1,0 +1,4 @@
+# TODO: define this as a singleton
+class UI:
+    def __init__(self):
+        self.tmp = ''
